refactor: new module fs_util
Will contain backward-compatible versions of filesystem-related functions.
